% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erlang.R
\name{erlang}
\alias{erlang}
\title{Utility function for mparse in SimInf}
\usage{
erlang(src, dst, shape, rate)
}
\arguments{
\item{src}{The name of the source compartment.}

\item{dst}{The name of the destination compartment.}

\item{shape}{The shape parameter (integer) of the Erlang
distribtion.}

\item{rate}{The rate parameter of the Erlang distribution.}
}
\description{
Create a model for an Erlang distribution in SimInf
}
\examples{
library(SimInf)
library(SWmisc)

shape <- 100
m <- erlang("A", "B", shape, rate = shape / 730)
u0 <- matrix(c(100, rep(0, length(m$compartments) - 1)),
             ncol = length(m$compartments))
u0 <- as.data.frame(u0)
colnames(u0) <- m$compartments
m <- mparse(m$transitions, m$compartments)
model <- init(m, u0, 1:1000)
plot(as.numeric(SimInf:::extract_U(run(model), "B")),
     type = "l", ylab = "Number of individuals in B compartment",
     xlab = "Time [day]")
}
