% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{file_stamp}
\alias{file_stamp}
\title{Create a time based stamp to be used as suffix in file names}
\usage{
file_stamp(time = "datetime", random = 0, sep = "-")
}
\arguments{
\item{time}{string with type of time string to use, "datetime" will output date and time (with only separator between date and date like 20220402-103402 for 2 May 2022, 10:34:02), default is "datetime"}

\item{random}{integer length of random string to add to the time, no random if 0 (default behavior)}

\item{sep}{string with separator to use between time and random part}
}
\description{
Create a time based stamp to be used as suffix in file names
}
