% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_r_file}
\alias{get_r_file}
\title{get file name with good extension .r or .R}
\usage{
get_r_file(file, should.exists = FALSE)
}
\arguments{
\item{file}{file path to find without extension}

\item{should.exists}{logical, error if file doesnt exists with any extension}
}
\value{
path
}
\description{
get file name with good extension .r or .R
}
