% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{has_package_version}
\alias{has_package_version}
\title{Compare installed version of a package}
\usage{
has_package_version(pkg, version, compare, not_found = NA)
}
\arguments{
\item{pkg}{package name}

\item{version}{version to test}

\item{compare}{comparator to use '>','<','==','>=','<='}

\item{not_found}{value to return if package is not installed}
}
\value{
boolean
}
\description{
Compare installed version of a package
}
