% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{merge_list}
\alias{merge_list}
\title{Merge two lists (x + y not in x)}
\usage{
merge_list(x, y)
}
\arguments{
\item{x}{primary list to update}

\item{y}{secondary list, only entries not in x will be used}
}
\description{
Missing entries in x will be taken from y. Dont handle nested lists (use \code{\link[utils]{modifyList}})
}
