% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{parse_int_list}
\alias{parse_int_list}
\title{Parse string to integer list accepting, comma separated values and range ('min-max') and 'all')}
\usage{
parse_int_list(v, acceptable.values, use.all = TRUE, unique = TRUE)
}
\arguments{
\item{v}{string to parse}

\item{acceptable.values}{list of acceptable values}

\item{use.all}{accept "all" special keyword, add values in acceptable.values}

\item{unique}{make values unique in list}
}
\description{
Parse string to integer list accepting, comma separated values and range ('min-max') and 'all')
}
