% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{parseArgs}
\alias{parseArgs}
\title{Parse command line arguments with the form var=value or var (no value=TRUE)}
\usage{
parseArgs(defs = NULL, args = NULL)
}
\arguments{
\item{defs}{arguments definition list of arguments (names = name of argument)}

\item{args}{argument list (used for test)}
}
\description{
Parse command line arguments.
If arguments definition is provided validate them
}
\details{
\describe{
\item{type}{"logical", "boolean","integer", "choices", can be abbreviated}
\item{required}{TRUE if argument is required}
\item{default}{Default value}
\item{description}{Description of option}
\item{choices}{acceptable choices, if type =="choices"}
\item{multiple}{if TRUE accept multiple values of the same type, comma separated}
\item{range}{if TRUE accept range, with the syntax from-to, can be comma separated }
\item{unique}{ensure uniqueness of values}
}
}
