% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{random_string}
\alias{random_string}
\title{Create a random sequence of n alphanumeric chars}
\usage{
random_string(n)
}
\arguments{
\item{n}{length of sequence}
}
\description{
Create a random sequence of n alphanumeric chars
}
