% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{replace_factor}
\alias{replace_factor}
\title{Replace factor levels by new values}
\usage{
replace_factor(.x, ..., .dots = NULL, .warn = TRUE)
}
\arguments{
\item{.x}{factor or vector of values (transformed into factor if not a factor)}

\item{...}{renaming argument new=old, NSE evaluation}

\item{.dots}{list with the form (new=old) of levels to}

\item{.warn}{if TRUE a warning will be thrown if unknown level is mentionned, else throw an error}
}
\description{
Replace factor levels by new values
}
