% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{replace_names}
\alias{replace_names}
\title{Replace names with given pairlist where value are old names and key are new names}
\usage{
replace_names(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{named list of column to rename (names are new names to assign)}
}
\description{
replace_names(data.frame(), new.name=old.name,...)
ex replace_names(d, toto="x") replace column "x" by "toto"
}
