% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbatch.R
\name{sbatch}
\alias{sbatch}
\title{Submit a job to slurm}
\usage{
sbatch(job = c("core", "node"), cores = 1, account = NULL,
  jobname = NULL, output = NULL, time = "10:00:00", script = NULL,
  filename = NULL)
}
\arguments{
\item{job}{What partition to choose.}

\item{cores}{The number of cores to use.}

\item{account}{Which project should be accounted for the running
time. If \code{account} is \code{NULL} the environment
variable \code{SLURM_ACCOUNT} is used.}

\item{jobname}{Name of job.}

\item{output}{Name of file for stdout.}

\item{time}{How long you want to reserve those nodes/cores
(Format: -t d-hh:mm:ss).}

\item{script}{Commands to run}

\item{filename}{Where to write the batch job script.}
}
\description{
Submit a job to slurm
}
