% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{select_df}
\alias{select_df}
\title{Select columns from a data.frame
Ancestor of dplyr::select()}
\usage{
select_df(data, columns, error = FALSE)
}
\arguments{
\item{data}{data.frame}

\item{columns}{list of column to extract}

\item{error}{raise error if unknown column}
}
\value{
data.frame
}
\description{
Select columns from a data.frame
Ancestor of dplyr::select()
}
