% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{wrap_string}
\alias{wrap_string}
\title{Add prefix and suffix to a string only if string is not empty}
\usage{
wrap_string(x, before = NULL, after = NULL)
}
\arguments{
\item{x}{string to wrap}

\item{before}{string to add as a suffix if x is not empty}

\item{after}{string to add as a postfix if x is not empty}
}
\description{
Add prefix and suffix to a string only if string is not empty
}
