% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erlang.R
\name{erlang}
\alias{erlang}
\title{Utility function for mparse in SimInf}
\usage{
erlang(src, dst, shape, rate)
}
\arguments{
\item{src}{The name of the source compartment.}

\item{dst}{The name of the destination compartment.}

\item{shape}{The shape parameter (integer) of the Erlang
distribtion.}

\item{rate}{The rate parameter of the Erlang distribution.}
}
\description{
Create a model for an Erlang distribution in SimInf
}
